+++ src/test/java/com/solace/spring/cloud/stream/binder/health/indicators/SolaceBinderHealthIndicatorTest.java
@@ -2,8 +2,8 @@ package com.solace.spring.cloud.stream.binder.health.indicators;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.actuate.health.Health;
-import org.springframework.boot.actuate.health.Status;
+import org.springframework.boot.health.contributor.Health;
+import org.springframework.boot.health.contributor.Status;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;